# swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  GenericError:
    properties:
      message:
        type: string
        x-go-name: Message
    type: object
    x-go-package: github.com/sssaang/go-cafe/api/handlers
  Product:
    description: Product defines the structure for an API product
    properties:
      description:
        description: the description for the product
        maxLength: 10000
        type: string
        x-go-name: Description
      id:
        description: the id for the user
        format: int64
        minimum: 1
        type: integer
        x-go-name: ID
      name:
        description: the name for the product
        maxLength: 255
        type: string
        x-go-name: Name
      price:
        description: the price for the product
        format: float
        minimum: 0.01
        type: number
        x-go-name: Price
      sku:
        description: the SKU for the product
        pattern: '[a-z]+-[a-z]+-[a-z]+'
        type: string
        x-go-name: SKU
    required:
    - id
    - name
    - price
    - sku
    type: object
    x-go-package: github.com/sssaang/go-cafe/api/data
  ValidationError:
    properties:
      message:
        type: string
        x-go-name: Message
    type: object
    x-go-package: github.com/sssaang/go-cafe/api/handlers
info:
  description: Documentation for Product API
  version: 1.0.0
paths:
  /products:
    get:
      description: Returns a list of all products
      operationId: listProducts
      responses:
        "200":
          $ref: '#/responses/productsResponse'
      tags:
      - products
    post:
      description: Create a new product
      operationId: createProduct
      parameters:
      - description: |-
          Product data structure to Update or Create.
          Note: the id field is ignored by update and create operations
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/Product'
      responses:
        "200":
          $ref: '#/responses/productResponse'
        "422":
          $ref: '#/responses/errorValidation'
        "501":
          $ref: '#/responses/errorResponse'
      tags:
      - products
  /products/{id}:
    delete:
      description: Deletes a product
      operationId: deleteProduct
      parameters:
      - description: id of the product to be deleted or modified from the database
        format: int64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "201":
          $ref: '#/responses/noContentResponse'
      tags:
      - products
    put:
      description: Update a products details
      operationId: updateProduct
      parameters:
      - description: id of the product to be deleted or modified from the database
        format: int64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      - description: |-
          Product data structure to Update or Create.
          Note: the id field is ignored by update and create operations
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/Product'
      responses:
        "201":
          $ref: '#/responses/noContentResponse'
        "404":
          $ref: '#/responses/errorResponse'
        "422":
          $ref: '#/responses/errorValidation'
      tags:
      - products
produces:
- application/json
responses:
  errorResponse:
    description: Generic error message returned as a string
    schema:
      $ref: '#/definitions/GenericError'
  errorValidation:
    description: Validation errors defined as an array of strings
    schema:
      $ref: '#/definitions/ValidationError'
  noContentResponse:
    description: A response with no content
  productResponse:
    description: Data structure representing a single product
    schema:
      $ref: '#/definitions/Product'
  productsResponse:
    description: A list of products returns in the response
    schema:
      items:
        $ref: '#/definitions/Product'
      type: array
schemes:
- http
swagger: "2.0"